move methods to rocketchat/methods
