+++ app/lib/rocketchat/methods/getRooms.js
