+++ app/lib/rocketchat/methods/helpers/findSubscriptionsRooms.js
@@ -1,6 +1,6 @@
 import { Q } from '@nozbe/watermelondb';
 
-import database from '../../database';
+import database from '../../../database';
 
 export default async (subscriptions = [], rooms = []) => {
 	try {
