+++ app/lib/rocketchat/methods/helpers/getFileUrlFromMessage.js
