+++ app/lib/rocketchat/methods/helpers/normalizeMessage.js
