+++ app/lib/rocketchat/methods/helpers/parseQuery.js
