+++ app/lib/rocketchat/methods/helpers/parseUrls.js
