+++ app/lib/rocketchat/methods/helpers/protectedFunction.js
