+++ app/lib/rocketchat/methods/loadMessagesForRoom.js
@@ -1,9 +1,9 @@
 import moment from 'moment';
 
-import { MESSAGE_TYPE_LOAD_MORE } from '../../constants/messageTypeLoad';
-import log from '../../utils/log';
-import { getMessageById } from '../database/services/Message';
-import { generateLoadMoreId } from '../utils';
+import { MESSAGE_TYPE_LOAD_MORE } from '../../../constants/messageTypeLoad';
+import log from '../../../utils/log';
+import { getMessageById } from '../../database/services/Message';
+import { generateLoadMoreId } from '../../utils';
 import updateMessages from './updateMessages';
 
 const COUNT = 50;
