+++ app/lib/rocketchat/methods/loadSurroundingMessages.js
@@ -2,10 +2,10 @@ import EJSON from 'ejson';
 import moment from 'moment';
 import orderBy from 'lodash/orderBy';
 
-import log from '../../utils/log';
-import { getMessageById } from '../database/services/Message';
-import { MESSAGE_TYPE_LOAD_NEXT_CHUNK, MESSAGE_TYPE_LOAD_PREVIOUS_CHUNK } from '../../constants/messageTypeLoad';
-import { generateLoadMoreId } from '../utils';
+import log from '../../../utils/log';
+import { getMessageById } from '../../database/services/Message';
+import { MESSAGE_TYPE_LOAD_NEXT_CHUNK, MESSAGE_TYPE_LOAD_PREVIOUS_CHUNK } from '../../../constants/messageTypeLoad';
+import { generateLoadMoreId } from '../../utils';
 import updateMessages from './updateMessages';
 
 const COUNT = 50;
