+++ app/lib/rocketchat/methods/loadThreadMessages.js
@@ -2,9 +2,9 @@ import { Q } from '@nozbe/watermelondb';
 import { sanitizedRaw } from '@nozbe/watermelondb/RawRecord';
 import EJSON from 'ejson';
 
-import database from '../database';
-import log from '../../utils/log';
-import { Encryption } from '../encryption';
+import database from '../../database';
+import log from '../../../utils/log';
+import { Encryption } from '../../encryption';
 import protectedFunction from './helpers/protectedFunction';
 import buildMessage from './helpers/buildMessage';
